+++ aiohttp_sqlalchemy/utils.py
@@ -5,7 +5,7 @@ from sqlalchemy import MetaData
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from aiohttp_sqlalchemy.constants import SA_DEFAULT_KEY
-from aiohttp_sqlalchemy.typedefs import TSessionFactory
+from aiohttp_sqlalchemy.typedefs import TOptSessionFactory
 
 
 async def init_db(
@@ -19,13 +19,14 @@ async def init_db(
     :param metadata: ...
     :param key: key of SQLAlchemy binding.
     """
-    session_factory = sa_session_factory(app, key)
-    async with session_factory() as session:
-        async with session.bind.begin() as connection:
-            await connection.run_sync(metadata.create_all)
+    session_factory = get_session_factory(app, key)
+    if callable(session_factory):
+        async with session_factory() as session:
+            async with session.bind.begin() as connection:
+                await connection.run_sync(metadata.create_all)
 
 
-def sa_session(
+def get_session(
     request: Request,
     key: str = SA_DEFAULT_KEY,
 ) -> AsyncSession:
@@ -45,17 +46,25 @@ def sa_session(
     return session
 
 
-def sa_session_factory(
+def get_session_factory(
     source: Union[Request, Application],
     key: str = SA_DEFAULT_KEY,
-) -> TSessionFactory:
+) -> TOptSessionFactory:
     """Return callable object which returns an `AsyncSession` instance.
 
     :param source: AIOHTTP request object or your AIOHTTP application.
     :param key: key of SQLAlchemy binding.
     """
-    return cast(TSessionFactory, getattr(source, "app", source).get(key))
+    if isinstance(source, Request):
+        return cast(TOptSessionFactory, source.config_dict.get(key))
+    elif isinstance(source, Application):
+        return cast(TOptSessionFactory, source.get(key))
+    raise TypeError(
+        'Arg `source` must be `Application` or `Request` from `aiohttp.web`.'
+    )
 
 
 # Synonyms
 sa_init_db = init_db
+sa_session = get_session
+sa_session_factory = get_session_factory