+++ docs/releases.rst
@@ -1,6 +1,18 @@
 ========
 Releases
 ========
+Version 0.19
+------------
+**Added**
+
+* Synonym ``sa_session`` for ``get_session``.
+* Synonym ``sa_session_factory`` for ``get_session_factory``.
+
+**Changed**
+
+* Rename ``sa_session`` to ``get_session``.
+* Rename ``sa_session_factory`` to ``get_session_factory``.
+
 Version 0.18.1
 --------------
 **Added**